ortho math_logstar_n: remove unnecessary cast
The RHS implicitly coerces to `double`.

+++ lib/ortho/trapezoid.c
@@ -851,7 +851,7 @@ static int math_logstar_n(int n)
   int i;
   double v;
 
-  for (i = 0, v = (double) n; v >= 1; i++)
+  for (i = 0, v = n; v >= 1; i++)
       v = log2(v);
 
   return i - 1;
